requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-slugify==8.0.1
requests==2.28.2
soupsieve==2.4
text-unidecode==1.3
urllib3==1.26.15
pylint==2.17.2
nltk==3.8.1
gensim==4.3.1
python-decouple==3.8
